Working A* code using the stl::priority queue.
There is a gotcha with the stl priority queue:  it is a max priority
queue, rather than a min priority queue.  Hence, the comparator needs
to take this into account.

This code appears to be faster than Ethan's code for Korf instances 1,
12, and 55---roughly 2x faster.

This code does not do any cleanup at the end, but rather drops
everything on the floor.  It seems a tad dishonest to publish results
this way, maybe, but calling the destructor on all the search nodes
could be expensive, and for this program, is not needed.

I bet I can make this code even faster.  Perhaps a special allocator
for TilesNode15 would help.