Moved BucketPriorityQueue; changed implementation significantly.
The new implementation requires that the Node template parameter have
methods get_g() and get_f().  So, it is more or less tied to
search/Node.hpp.

Rather than being implemented as a vector of vectors of nodes, the new
implementation uses a vector of vectors of lists of nodes.  All the
uses of stl heap algorithms are gone.  Instead, the element type is
binned based on its f-value, and then binned by its g-value.  This
allows for (what should be) truly constant time priority queue
operations, ordering nodes by f, and breaking ties in favor of high
g.  The old implementation had some logarithmic complexity for the
pop/top/push operations, due to the use of the heap algorithms.