Massive amounts of debugging asserts / debugging output.
This all started in an attempt to allow fast deletion from the open
list.  The idea is this: instead of keeping a closed set in the A*
search, a closed mapping is kept instead.  (``Closed'' is a bad name,
since the mapping keeps track of all nodes that have ever been
expanded or that are in the open list.)  An entry in the closed set
can possibly map to an object (essentially a pointer?) pointing into
the open list, which can be given as an argument to an ``erase''
method of the priority queue.

I have been chasing memory-related bugs for hours.  Here is some
background information on the problem:

  * The priority queue is implemented with a vector of vectors of
    lists of Node pointers.
  * When inserting a node into the priority queue, a pointer to its
    location in the open list is returned.  These returned pointers
    are stuffed into the closed list.
  * When generating nodes, if a node has been closed, it is pruned.
    Otherwise, if a copy of the node is present in the open list and
    the new copy has a lower f-value, the old copy is removed using
    the ``erase'' method, and the new copy is inserted into both the
    open list and closed list.

I am noticing memory corruption after several nodes have been
generated/inserted/erased/etc, i.e., after the A* loop has completed a
few times.

What I think is happening is this: the priority queue's internal
vectors eventually grow too large, and need to be resized.  If the
vectors' stores cannot simply be enlarged, a new, larger store must be
allocated, and the old contents copied over.  When this occurs, all
the pointers into the open list that the priority queue handed out
have been made invalid.

A fix awaits!