Yet another rewrite of the bucket-based priority queue.  Mostly working.
All but the erase operation appears to work.  To implement fast
(i.e. amortized constant time) erase, the pop & top methods will need
to be modified as well.  I erase elements given an ItemPointer by
setting the pointed-to-location to NULL.

I doubt this is the best way to implement amortized constant-time
push, pop, top, and delete, but it seems reasonable.