Hash table initial size hint; comments.
In the previous commit log, I wrote that this data structure is slow.
It is slower than I would like it to be, but I had forgotten that I
deleted the initial closed set size hint during a debugging session.

Adding the size hint back in, the performance on Korf #1 is down to
~34 seconds, from 50-something before.  Resizing hash tables can be
expensive!