Rewrote abstraction code for efficiency.
Before, I would do a linear scan, for each tile in a TilesState15, to
determine if the tile should be abstracted at the given level.  This
revision implements abstraction by constructing an 2d array of
booleans, which allows abstraction to work by walking down the tiles
array and doing a single array lookup for each element.