Separated heuristic calculation from node expansion; renamed & added methods.
`TilesInstance15::expand' has been renamed to
TilesInstance15::compute_successors.

I also added TilesInstance15::compute_predecessors.  This is simply an
alias for compute_successors in this domain, as the actions are all
reversible.

Splitting heuristic calculation from node generation is necessary for
Switchback, which computes a pattern database on the fly.