A* bugfixes?
I was freeing old copies of nodes in the open list in a strange order.

open_ptr was a non-const reference to the closed list's iterator's
second.  I later changed the value of open_ptr.  This was probably a
nasty bug waiting to happen!

I was not freeing pruned successor nodes in all cases.  Fixed.