Working switchback!?!?
I was getting suboptimal solutions to certain tiles problems.  I think
the reason for this is because my ``closed'' list actually keeps track
of everything that is properly closed, as well as everything that is
in open.  So ``closed''---a misnomer---holds on to exactly one copy of
every state that has ever been generated.

I was writing my code that checks for cached values (i.e., in
resume_search and heuristic) as if closed were only keeping track of
truly closed nodes.

Instead of simply checking if there is an entry in the closed set for
a node to determine if the cached result should be used, we need to
also check if the cached node is not in the open list.  This is a
one-line fix in two places.

I also rewrote the loop in init_open_and_closed, to count up.  A count
up loop is cleaner, and it is all that is necessary, since a heuristic
value of 0 is used for each initially inserted node.