Do not call destructors on any of the top-level objects.
This is kind of nasty, and is a memory leak, but it doesn't matter.
By not calling the destructors on the objects, the program executes
much faster.  The OS does a fine job of cleaning up the memory.  Since
I am not using any resources that require explicit cleanup for the
computer to function properly (such as sockets), this approach should
be fine.