Added preprocessor flag:  CACHE_TILES_HASH_VALUE
When this is not defined, a tile puzzle's hash value is calculated
as-needed.  This helps memory usage some.  I'm not sure how
significant the speed impact is; on betelgeuse, the timings are very
inconsistent from run to run.  It seems a tad slower when
CACHE_TILES_HASH_VALUE is not defined.