Working (?) implementation of naive hierarchical A*.
I think it is working, but it is horrifically slow, as expected.  It
has been running for ~40 minutes on my desktop, and has generated over
3 billion nodes, but only has expanded about 2200 nodes at the base
level.  It's going to be a while!

Actually, though, this is not quite a correct implementation of naive
hierarchical A*.  In the HA* tech report, they say that they use the
``min operator cost'' heuristic as a default, and take the maximum.  I
do not do this.