Made the number of abstraction levels easily modifiable.
To create a custom abstraction of n levels, obscure the first 15 - n
tiles at the first level, and a single tile after that.

Although my change reads (14 - num_abstraction_levels), it is really
the same as 15 - n, due to the addition of `level'.

It turns out that num_abstraction_levels of 7 results in fewer node
expansions with Switchback on Korf #1, #12, and #55.