Smarter P-g caching code.
Rather than scanning over the entire closed list for a level (which
contains a single copy of every state generated during the search),
when expanding nodes, I add them to a vector of expanded nodes.  Then,
to do P-g caching, only this vector of nodes needs to be iterated
over.

This tweak brings HA* run time from about 2:25 on Korf 13 down to about
2:07, or about a 12% speed increase.