Bugfix in `search' procedure in driver; added support for more algorithms with Pancakes.
`search' was a template procedure, but used TileNode15 in its body.
This is not the node type for the pancake puzzles.

I changed line 188 in src/Search.cpp so that it is properly generic,
using the Node type of the Searcher template parameter.  Now, a found
goal node is printed to cout.

To make this work, a public Node typedef had to be added to each of
the search algorithms.

I also did some small refactoring in Search.cpp, such as renaming
variables.

I added the option of running the pancake puzzle using IDA* and HA*.